ACMS-3360: Refactor sitestudio_config_management module.
